+++ src/main/java/chess/WebApplication.java
@@ -3,7 +3,7 @@ package chess;
 import static spark.Spark.port;
 import static spark.Spark.staticFileLocation;
 
-import chess.web.GameController;
+import chess.controller.GameController;
 
 public class WebApplication {
     public static void main(String[] args) {
