+++ src/main/java/chess/dao/PieceDao.java
@@ -1,4 +1,4 @@
-package chess.web.dao;
+package chess.dao;
 
 import chess.domain.board.Piece;
 import chess.domain.position.Position;
