+++ src/main/java/chess/dao/PieceDaoImpl.java
@@ -1,4 +1,4 @@
-package chess.web.dao;
+package chess.dao;
 
 import chess.domain.board.Color;
 import chess.domain.board.Piece;
