+++ src/main/java/chess/dto/MoveDto.java
@@ -1,4 +1,4 @@
-package chess.web.dto;
+package chess.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
