+++ src/main/java/chess/view/command/Status.java
@@ -1,13 +1,12 @@
-package chess.console.command;
+package chess.view.command;
 
-import chess.console.view.OutputView;
+import chess.view.OutputView;
 import chess.domain.GameManager;
 
-public final class Start implements Command {
-
+public final class Status implements Command {
     @Override
     public void execute(GameManager gameManager) {
-        OutputView.printBoard(gameManager.getBoard());
+        OutputView.printStatus(gameManager.getStatus());
     }
 
     @Override
