+++ src/test/java/chess/dao/TurnDaoTest.java
@@ -3,7 +3,6 @@ package chess.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.state.Turn;
-import chess.web.dao.TurnDaoImpl;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
